refactor(DeckTest): GameStatisticsTest 내의 JudgeTest를 ParticipantTest로 분리
